Migrate history to the environement object
Fully removes the pipelines from having to keep track of history. This
will enable us to move to a batch or multi agent setup easier.
