+++ bindsnet/environment/__init__.py
@@ -5,7 +5,7 @@ import numpy as np
 from typing import Tuple, Dict, Any
 from abc import ABC, abstractmethod
 
-from ..datasets import MNIST, CIFAR10, CIFAR100, SpokenMNIST
+from ..datasets.spike_encoders import Encoder
 from ..datasets.preprocess import subsample, gray_scale, binary_image, crop
 
 
@@ -57,14 +57,6 @@ class Environment(ABC):
         """
         pass
 
-    @abstractmethod
-    def reshape(self) -> None:
-        # language=rst
-        """
-        Abstract method header for ``reshape()``.
-        """
-        pass
-
 
 class GymEnvironment(Environment):
     # language=rst
@@ -72,25 +64,41 @@ class GymEnvironment(Environment):
     A wrapper around the OpenAI ``gym`` environments.
     """
 
-    def __init__(self, name: str, **kwargs) -> None:
+    def __init__(self, name: str, encoder: Encoder, **kwargs) -> None:
         # language=rst
         """
         Initializes the environment wrapper.
 
         :param name: The name of an OpenAI :code:`gym` environment.
+        :param encoder: Function to encode observations into spike trains.
 
         Keyword arguments:
 
         :param max_prob: Maximum spiking probability.
         :param clip_rewards: Whether or not to use :code:`np.sign` of rewards.
+
+        :param int history: Number of observations to keep track of.
+        :param int delta: Step size to save observations in history.
         """
         self.name = name
         self.env = gym.make(name)
         self.action_space = self.env.action_space
+        self.encoder = encoder
 
         # Keyword arguments.
         self.max_prob = kwargs.get("max_prob", 1)
         self.clip_rewards = kwargs.get("clip_rewards", True)
+
+        self.history_length = kwargs.get('history_length', None)
+        self.delta = kwargs.get('delta', 1)
+
+        if self.history_length is not None and self.delta is not None:
+            self.history = {i: torch.Tensor() for i in range(1, self.history_length * self.delta + 1, self.delta)}
+        else:
+            self.history = {}
+
+        self.episode_step_count = 0
+        self.history_index = 1
 
         self.obs = None
         self.reward = None
@@ -113,6 +121,15 @@ class GymEnvironment(Environment):
 
         self.preprocess()
 
+        # Store frame of history and encode the inputs.
+        if len(self.history) > 0:
+            self.update_history()
+            self.update_index()
+
+        self.obs = self.encoder(self.obs)
+
+        self.episode_step_count += 1
+
         # Return converted observations and other information.
         return self.obs, self.reward, self.done, info
 
@@ -126,6 +143,10 @@ class GymEnvironment(Environment):
         # Call gym's environment reset function.
         self.obs = self.env.reset()
         self.preprocess()
+
+        self.history = {i: torch.Tensor() for i in self.history}
+
+        self.episode_step_count = 0
 
         return self.obs
 
@@ -161,11 +182,39 @@ class GymEnvironment(Environment):
 
         self.obs = torch.from_numpy(self.obs).float()
 
-    def reshape(self) -> torch.Tensor:
+    def update_history(self) -> None:
         # language=rst
         """
-        Reshape observation for plotting purposes.
-
-        :return: Reshaped observation to plot in ``plt.imshow()`` call.
+        Updates the observations inside history by performing subtraction from  most recent observation and the sum of
+        previous observations. If there are not enough observations to take a difference from, simply store the
+        observation without any differencing.
         """
-        return self.obs
+        # Recording initial observations
+        if self.episode_step_count < len(self.history) * self.delta:
+            # Store observation based on delta value
+            if self.episode_step_count % self.delta == 0:
+                self.history[self.history_index] = self.obs
+        else:
+            # Take difference between stored frames and current frame
+            temp = torch.clamp(self.obs - sum(self.history.values()), 0, 1)
+
+            # Store observation based on delta value.
+            if self.episode_step_count % self.delta == 0:
+                self.history[self.history_index] = self.obs
+
+            assert (len(self.history) == self.history_length), 'History size is out of bounds'
+            self.obs = temp
+
+    def update_index(self) -> None:
+        # language=rst
+        """
+        Updates the index to keep track of history. For example: history = 4, delta = 3 will produce self.history = {1,
+        4, 7, 10} and self.history_index will be updated according to self.delta and will wrap around the history
+        dictionary.
+        """
+        if self.episode_step_count % self.delta == 0:
+            if self.history_index != max(self.history.keys()):
+                self.history_index += self.delta
+            else:
+                # Wrap around the history.
+                self.history_index = (self.history_index % max(self.history.keys())) + 1
