+++ bindsnet/pipeline/environment_pipeline.py
@@ -23,23 +23,17 @@ class RLPipeline(BasePipeline):
     """
 
     def __init__(self, network: Network, environment: Environment, encoding: Callable = bernoulli,
-                 action_function: Optional[Callable] = None, enable_history: Optional[bool] = False,
-                 **kwargs):
+                 action_function: Optional[Callable] = None, **kwargs):
         # language=rst
         """
         Initializes the pipeline.
 
         :param network: Arbitrary network object.
         :param environment: Arbitrary environment.
-        :param encoding: Function to encode observations into spike trains.
         :param action_function: Function to convert network outputs into environment inputs.
-        :param enable_history: Enable history functionality.
 
         Keyword arguments:
 
-        :param int time: Time input is presented for to the network.
-        :param int history: Number of observations to keep track of.
-        :param int delta: Step size to save observations in history.
         :param int render_interval: Interval to render the environment.
 
         :param str output: String name of the layer from which to take output from.
@@ -50,40 +44,26 @@ class RLPipeline(BasePipeline):
         super().__init__(network, **kwargs)
 
         self.env = environment
-        self.encoding = encoding
         self.action_function = action_function
-        self.enable_history = enable_history
 
-        self.history_index = 1
         self.accumulated_reward = 0.0
         self.reward_list = []
 
         # Setting kwargs.
-        self.time = kwargs.get('time', 1)
-        self.delta = kwargs.get('delta', 1)
         self.output = kwargs.get('output', None)
-        self.history_length = kwargs.get('history_length', None)
         self.render_interval = kwargs.get('render_interval', None)
         self.reward_window = kwargs.get('reward_window', None)
         self.reward_delay = kwargs.get('reward_delay', None)
         self.num_episodes = kwargs.get('num_episodes', 100)
-
-        self.dt = network.dt
-        self.timestep = int(self.time / self.dt)
-
-        if self.history_length is not None and self.delta is not None:
-            self.history = {i: torch.Tensor() for i in range(1, self.history_length * self.delta + 1, self.delta)}
-        else:
-            self.history = {}
 
         if self.reward_delay is not None:
             assert self.reward_delay > 0
             self.rewards = torch.zeros(self.reward_delay)
 
         # Set up for multiple layers of input layers.
-        self.encoded = {
-            name: torch.Tensor() for name, layer in network.layers.items() if isinstance(layer, AbstractInput)
-        }
+        self.inpts = [
+            name for name, layer in network.layers.items() if isinstance(layer, AbstractInput)
+        ]
 
         self.action = None
         self.obs = None
@@ -133,18 +113,8 @@ class RLPipeline(BasePipeline):
         # Accumulate reward
         self.accumulated_reward += self.reward
 
-        # Store frame of history and encode the inputs.
-        if self.enable_history and len(self.history) > 0:
-            self.update_history()
-            self.update_index()
-
-        # Encode the observation using given encoding function.
-        for inpt in self.encoded:
-            self.encoded[inpt] = self.encoding(self.obs, time=self.time, dt=self.network.dt)
-
         return {'obs': self.obs, 'reward': reward,
-                'done': self.done, 'info': info,
-                'encoded_obs': self.encoded}
+                'done': self.done, 'info': info}
 
     def step_(self, batch) -> None:
         # language=rst
@@ -152,55 +122,17 @@ class RLPipeline(BasePipeline):
         Run an iteration of the network and log any needed data
         """
 
-        encoded = batch['encoded_obs']
+        inpts = {k: batch['obs'] for k in self.inpts}
         reward = batch['reward']
 
         # Run the network on the spike train-encoded inputs.
-        self.network.run(inpts=encoded, time=self.time, reward=reward)
+        self.network.run(inpts=inpts, time=batch['obs'].shape[0], reward=reward)
 
         if self.done:
             if self.network.reward_fn is not None:
                 self.network.reward_fn.update(**kwargs)
-            self.step_count = 0
             self.reward_list.append(self.accumulated_reward)
             self.accumulated_reward = 0.0
-
-    def update_history(self) -> None:
-        # language=rst
-        """
-        Updates the observations inside history by performing subtraction from  most recent observation and the sum of
-        previous observations. If there are not enough observations to take a difference from, simply store the
-        observation without any differencing.
-        """
-        # Recording initial observations
-        if self.step_count < len(self.history) * self.delta:
-            # Store observation based on delta value
-            if self.step_count % self.delta == 0:
-                self.history[self.history_index] = self.obs
-        else:
-            # Take difference between stored frames and current frame
-            temp = torch.clamp(self.obs - sum(self.history.values()), 0, 1)
-
-            # Store observation based on delta value.
-            if self.step_count % self.delta == 0:
-                self.history[self.history_index] = self.obs
-
-            assert (len(self.history) == self.history_length), 'History size is out of bounds'
-            self.obs = temp
-
-    def update_index(self) -> None:
-        # language=rst
-        """
-        Updates the index to keep track of history. For example: history = 4, delta = 3 will produce self.history = {1,
-        4, 7, 10} and self.history_index will be updated according to self.delta and will wrap around the history
-        dictionary.
-        """
-        if self.step_count % self.delta == 0:
-            if self.history_index != max(self.history.keys()):
-                self.history_index += self.delta
-            else:
-                # Wrap around the history.
-                self.history_index = (self.history_index % max(self.history.keys())) + 1
 
     def reset_(self) -> None:
         # language=rst
@@ -209,9 +141,7 @@ class RLPipeline(BasePipeline):
         """
         self.env.reset()
         self.network.reset_()
-        self.step_count = 0
         self.accumulated_reward = 0.0
-        self.history = {i: torch.Tensor() for i in self.history}
 
     def plots(self, batch, *args):
         return
