+++ examples/space_invaders/minimal_space_invaders.py
@@ -3,7 +3,7 @@ import torch
 from bindsnet.network import Network
 from bindsnet.pipeline import RLPipeline
 from bindsnet.learning import MSTDPET
-from bindsnet.encoding import bernoulli
+from bindsnet.datasets.spike_encoders import PoissonEncoder
 from bindsnet.network.topology import Connection
 from bindsnet.environment import GymEnvironment
 from bindsnet.network.nodes import Input, LIFNodes
@@ -36,13 +36,14 @@ network.add_connection(inpt_middle, source="X", target="Y")
 network.add_connection(middle_out, source="Y", target="Z")
 
 # Load SpaceInvaders environment.
-environment = GymEnvironment("SpaceInvaders-v0")
+environment = GymEnvironment('SpaceInvaders-v0',
+                             PoissonEncoder(time=1, dt=network.dt),
+                             history_length=2, delta=4)
 environment.reset()
 
 # Build pipeline from specified components.
-pipeline = RLPipeline(network, environment, encoding=bernoulli,
+pipeline = RLPipeline(network, environment,
                     action_function=select_multinomial, output='Z',
-                    time=1, history_length=2, delta=4,
                     plot_interval=100, render_interval=5)
 
 # Run environment simulation and network training.
